requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
advanced-alchemy==1.4.3
    # via -r requirements.in
alembic==1.15.2
    # via advanced-alchemy
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via starlette
asyncpg==0.30.0
    # via -r requirements.in
bcrypt==4.0.1
    # via
    #   -r requirements.in
    #   passlib
black==25.1.0
    # via -r requirements.in
click==8.1.8
    # via
    #   black
    #   uvicorn
fastapi==0.115.12
    # via -r requirements.in
greenlet==3.2.2
    # via advanced-alchemy
h11==0.16.0
    # via uvicorn
idna==3.10
    # via anyio
iniconfig==2.1.0
    # via pytest
isort==6.0.1
    # via -r requirements.in
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via mako
mypy-extensions==1.1.0
    # via black
packaging==25.0
    # via
    #   black
    #   pytest
passlib[bcrypt]==1.7.4
    # via -r requirements.in
pathspec==0.12.1
    # via black
platformdirs==4.3.8
    # via black
pluggy==1.5.0
    # via pytest
pydantic==2.11.4
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.9.1
    # via -r requirements.in
pyjwt==2.10.1
    # via -r requirements.in
pytest==8.3.5
    # via -r requirements.in
python-dotenv==1.1.0
    # via pydantic-settings
ruff==0.11.8
    # via -r requirements.in
scalar-fastapi==1.0.3
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.41
    # via
    #   -r requirements.in
    #   advanced-alchemy
    #   alembic
starlette==0.46.2
    # via fastapi
typing-extensions==4.13.2
    # via
    #   advanced-alchemy
    #   alembic
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.0
    # via
    #   pydantic
    #   pydantic-settings
uvicorn==0.34.2
    # via -r requirements.in
